Exercises 4.15.-4.23.

+++ part4/blog_list/controllers/blogs.js
@@ -1,9 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/Blogs')
+const { userExtractor } = require('../utils/middleware')
 
-blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({})
-  res.json(blogs)
+
+blogsRouter.get('/', async (req, res, next) => {
+  try {
+    const blogs = await Blog.find({})
+      .populate('user', { username: 1, name: 1 })
+    res.json(blogs)
+  } catch (err) {
+    next(err)
+  }
 })
 
 
@@ -20,25 +27,36 @@ blogsRouter.get('/:id', async (req, res, next) => {
   }
 })
 
-blogsRouter.post('/', async (req, res, next) => {
+blogsRouter.post('/', userExtractor, async (req, res, next) => {
   const body = req.body
-  const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes,
-  })
   try {
+    const user = req.user
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id,
+    })
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
     res.status(201).json(savedBlog)
   } catch (err) {
     next(err)
   }
 })
 
-blogsRouter.delete('/:id', async (req, res, next) => {
+blogsRouter.delete('/:id', userExtractor, async (req, res, next) => {
   try {
+    const user = req.user
+    const blog = await Blog.findById(req.params.id)
+    if (blog && blog.user.toString() !== user._id.toString() ) {
+      return res.status(403).json({ error: 'forbidden to delete blog' })
+    }
     await Blog.findByIdAndDelete(req.params.id)
+    user.blogs = user.blogs.filter(b => b.toString() !== blog._id.toString())
+    await user.save()
     res.status(204).end()
   } catch (error) {
     next(error)
